Add diagrams from draw.io
I'm trying to figure out how best to share these in such a way that we
can all contribute to them easily. While integration between draw.io
and GitHub exists, it requires read/write permissions to all ParityTech
repos, which doesn't comfort me. I think having the .drawio file and
the corresponding pictures available in a folder is a good start.

I imagine the workflow would be something like this: Someone wants to
update a diagram, so they go to draw.io and upload the .drawio file
from the repo. They make their changes, download the new .drawio file
and corresponding SVG, and commit their changes to the repo.